+++ src/cpp_bindings/solver/config_solver.cpp
@@ -13,25 +13,37 @@
 
 namespace py = pybind11;
 
+// TODO: This does not need the template. Either binding type_descriptor or
+// accept the dtype= like numpy array.
+template <typename ValueType>
+std::shared_ptr<gko::LinOp> config_solver(std::shared_ptr<gko::Executor> exec,
+                                          std::shared_ptr<gko::LinOp> A,
+                                          std::string json)
+{
+    auto config = gko::ext::config::parse_json(nlohmann::json::parse(json));
+    // Create the registry, which allows passing the existing data into config
+    // This example does not use existing data.
+    auto reg = gko::config::registry();
+    // generate the linopfactory on the given executors
+    auto solver_gen =
+        gko::config::parse(config, reg,
+                           gko::config::make_type_descriptor<ValueType>())
+            .on(exec);
+
+    // Create solver
+    auto solver = solver_gen->generate(A);
+
+    return solver;
+}
+
 template <typename ValueType>
 std::shared_ptr<const gko::log::Convergence<ValueType>> config_solve(
     std::shared_ptr<gko::Executor> exec, std::shared_ptr<gko::LinOp> A,
     std::shared_ptr<gko::LinOp> b, std::shared_ptr<gko::LinOp> x,
     std::string json)
 {
-    auto config = gko::ext::config::parse_json(nlohmann::json::parse(json));
-    // Create the registry, which allows passing the existing data into config
-    // This example does not use existing data.
-    auto reg = gko::config::registry();
-    // Create the default type descriptor, which gives the default common type
-    // (value/index) for solver generation. If the solver does not specify value
-    // type, the solver will use these types.
-    auto td = gko::config::make_type_descriptor<ValueType, IndexType>();
-    // generate the linopfactory on the given executors
-    auto solver_gen = gko::config::parse(config, reg, td).on(exec);
-
     // Create solver
-    auto solver = solver_gen->generate(A);
+    auto solver = config_solver<ValueType>(exec, A, json);
 
     // Add logger
     std::shared_ptr<const gko::log::Convergence<ValueType>> logger =
@@ -49,6 +61,9 @@ void init_config_solver(py::module_ &m, const std::string value_type)
     std::string pyfunc_name = "config_solve_" + value_type;
     m.def(pyfunc_name.c_str(), &config_solve<ValueType>,
           "wrapper for config solve");
+    pyfunc_name = "config_solver_" + value_type;
+    m.def(pyfunc_name.c_str(), &config_solver<ValueType>,
+          "wrapper for generating config solver");
 }
 
 void init_config_solver_all_types(py::module_ &m)
@@ -56,4 +71,5 @@ void init_config_solver_all_types(py::module_ &m)
 #define DECLARE_CONFIG_SOLVER(ValueType) \
     init_config_solver<ValueType>(m, #ValueType);
     PYGKO_INSTANTIATE_FOR_EACH_NON_COMPLEX_VALUE_TYPE(DECLARE_CONFIG_SOLVER);
+#undef DECLARE_CONFIG_SOLVER
 }